requirements.txt
streamlit
requests
python-docx
pypdf
dotenv


